streamlit-automl/requirements.txt
streamlit~=1.34.0
snowflake-ml-python==1.4.0
python-dotenv
xgboost==1.7.3
scikit-learn==1.3.0
pandas~=1.5.3
altair~=5.3.0
black~=24.4.2
nbformat~=5.10.4
nbconvert~=7.16.4
numpy~=1.26.3
